Remove #reset_otp_state hook
This feature was added so enable and unconfirmed mobile
phone number to be used for OTP purposes and have that
number removed at the start or end of a new session.

However, a simpler way do this would be to simply store
the unconfirmed mobile in the #user_session.  Indeed
that is what we are now doing in 18F identity project
which uses this gem:
https://github.com/18F/identity-idp/pull/220