Adds options hash to method call on resource
**Why**
The only way to configure OTP delivery options is by managing state with
attributes. This adds more flexibility in implementating the
:send_two_factor_authentication_code method.

**How**
Have the gem pass the options hash back to the send_two_factor_authentication_code
method on the resource. This permits at run-time options to reside with the send_new_otp
call. This also maintains compatibility with existing implementations where
:send_two_factor_authentication_code only accepts one argument.