Make totp_timestamp column an integer
The ROTP gem expects the prior timestamp in
`verify_with_drift_and_prior` to be an integer. Moreover, when
`verify_with_drift_and_prior` returns it returns an integer.

We noticed a bug where the totp_timestamp column was always nil. The
reason for this was that the database had a timestamp column and when
this gem was attempting to assign `totp_timestamp` to an integer, it
was actually set to nil.

This commit changes the column to an integer instead of an integer to
avoid the problem described above.